+++ spec/resources.rb
@@ -1,0 +1,8 @@
+def resource(filename)
+  File.expand_path("../resources/#{filename}", __FILE__)
+end
+module Opower::TimeSeries::Test
+  # Test Resources
+  module Resources
+  end
+end