+++ spec/spec_helper.rb
@@ -1,12 +1,15 @@
-# -*- encoding: utf-8 -*-
-
-require 'rspec'
 require 'simplecov'
 
 SimpleCov.start 'rails' do
   coverage_dir 'metrics/coverage'
 end
 
-RSpec.configure do |config|
-  config.mock_with :mocha
+pid = Process.pid
+SimpleCov.at_exit do
+  SimpleCov.result.format! if Process.pid == pid
 end
+
+$LOAD_PATH << File.expand_path('../../lib', __FILE__)
+
+require File.expand_path('../fixtures.rb', __FILE__)
+require File.expand_path('../resources.rb', __FILE__)